chore: improve tree shaking
