+++ src/index.ts
@@ -1,4 +1,10 @@
 export { addV, subV } from './utils/math'
 export { rubberbandIfOutOfBounds } from './utils/rubberband'
 
-export * from './hooks'
+export { useDrag } from './hooks/useDrag'
+export { usePinch } from './hooks/usePinch'
+export { useWheel } from './hooks/useWheel'
+export { useMove } from './hooks/useMove'
+export { useHover } from './hooks/useHover'
+export { useScroll } from './hooks/useScroll'
+export { useGesture } from './hooks/useGesture'
